melhorando performance header
